Conflicts:
	Crocante/src/main/java/com/unal/crocante/LocalProveedor.java

+++ Crocante/src/main/java/com/unal/crocante/venta/EditarVenta.java
@@ -207,13 +207,13 @@ public class EditarVenta extends javax.swing.JFrame {
         String formaPago = null;
         String estadoPago = null;
         String modalidad = null;
-        int id = 0;
+        long id = 0;
         try {
             precio = Integer.parseInt(precioFormattedTextField.getText());
             formaPago = formaPagoComboBox.getSelectedItem().toString();
             estadoPago = estadoPagoComboBox.getSelectedItem().toString();
             modalidad = modalidadComboBox.getSelectedItem().toString();
-            id = Integer.parseInt(idFormattedTextField.getText());
+            id = Long.parseLong(idFormattedTextField.getText());
 
         } catch (NumberFormatException numberFormatException) {
             Logger.getLogger(AgregarVenta.class.getName()).log(Level.SEVERE, null, numberFormatException);
@@ -337,7 +337,7 @@ public class EditarVenta extends javax.swing.JFrame {
             formaPagoComboBox.setSelectedItem(resultado.getString(4));
             estadoPagoComboBox.setSelectedItem(resultado.getString(5));
             modalidadComboBox.setSelectedItem(resultado.getString(6));
-            idFormattedTextField.setValue(resultado.getInt(8));
+            idFormattedTextField.setValue(resultado.getLong(8));
 
         } catch (SQLException ex) {
             Logger.getLogger(VentaPrincipal.class.getName()).log(Level.SEVERE, null, ex);
